backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0